excited for first props warning/error! also,
the terminology gets confusing but i'm adding a bit of view data to this view.

that is,

the user is going to only want to view some of the data. the state
of the view is also perfectly good data. whether this is two signal
graphs or one will be left to future MEs to decide.