I'm thinking I need to improve my data model now.
I'm doing a bunch of stupid things in order to massage my data from
my views. So that can't be right, right? Yet I don't want to have
to write a server yet, not really.

So, I think that now that you have shitty but visible material-ui,
it's time to go back to the data model. Use the (:require db) trick
just to form the data into a shape that you can manage, and try
working with datascript in the app, terrifying though that may be.